requirements_saas.txt
# Virsaas Virtual Software Inc. - SaaS Platform Requirements
# Complete business platform with user accounts, subscriptions, and persistent simulations

# Web Framework
flask>=2.3.0
flask-sqlalchemy>=3.0.0
flask-login>=0.6.0

# Database
sqlalchemy>=2.0.0

# Authentication & Security
werkzeug>=2.3.0

# Payment Processing
stripe>=6.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# GUI and Visualization
pygame-ce>=2.3.0
pygame_gui>=0.6.0
plotly>=5.15.0

# Utilities
python-dateutil>=2.8.0
requests>=2.31.0
psutil>=5.9.0

# Development Tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Optional for enhanced features
# pywin32>=306  # Windows-specific
# winshell>=0.6.2  # Windows-specific

# Updated package requirements:
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
gunicorn==21.2.0
gevent==23.9.1
python-dotenv==1.0.0
# (any other deps you import)

#changed pages and features
cryptography==41.0.7
PyJWT==2.8.0
redis==5.0.1
httpx==0.25.2

#payment gateway setup
stripe==7.9.0
python-dateutil==2.8.2

#enhanced coding project features
flask-mail==0.9.1
PyGithub==2.1.1

#websocket support
flask-socketio==5.3.6
python-socketio==5.10.0
eventlet==0.33.3  # async mode

#improve whisper with ceo ai
openai-whisper==20231117
pydub==0.25.1

#misisng definintions from recent edits
cryptography==41.0.7
openai==0.28.1
solana==0.29.3              # Phantom wallet
base58==2.1.1               # Solana addressing